:truck: Move home page components to sections
