+++ src/components/index.js
@@ -1,49 +1,25 @@
-import About from './about';
-import Author from './author';
-import Blog from './blog';
-import Comments from './comments';
-import Contact from './contact';
-import DetailsAndSummary from './detailsAndSummary';
-import Education from './education';
-import Email from './email';
-import ExternalLink from './externalLink';
-import Featured from './featured';
-import Footer from './footer';
-import Head from './head';
-import Hero from './hero';
-import Jobs from './jobs';
-import Layout from './layout';
-import Loader from './loader';
-import Menu from './menu';
-import Nav from './nav';
-import Notifications from './notifications';
-import NowPlaying from './now-palying';
-import Projects from './projects';
-import Social from './social';
-import SocialShare from './socialShare';
+export { default as Author } from './author';
+export { default as Comments } from './comments';
+export { default as DetailsAndSummary } from './detailsAndSummary';
+export { default as Email } from './email';
+export { default as ExternalLink } from './externalLink';
+export { default as Footer } from './footer';
+export { default as Head } from './head';
+export { default as Layout } from './layout';
+export { default as Loader } from './loader';
+export { default as Menu } from './menu';
+export { default as Nav } from './nav';
+export { default as Notifications } from './notifications';
+export { default as NowPlaying } from './now-palying';
+// export { default as Side } from './side';
+export { default as Social } from './social';
+export { default as SocialShare } from './socialShare';
 
-export {
-  About,
-  Author,
-  Blog,
-  Comments,
-  Contact,
-  DetailsAndSummary,
-  Education,
-  Email,
-  ExternalLink,
-  Featured,
-  Footer,
-  Head,
-  Hero,
-  Jobs,
-  Layout,
-  Loader,
-  Menu,
-  Nav,
-  Notifications,
-  NowPlaying,
-  Projects,
-  Social,
-  SocialShare,
-};
+export { default as Hero } from './sections/hero';
+export { default as About } from './sections/about';
+export { default as Jobs } from './sections/jobs';
+export { default as Featured } from './sections/featured';
+export { default as Projects } from './sections/projects';
+export { default as Blog } from './sections/blog';
+export { default as Education } from './sections/education';
+export { default as Contact } from './sections/contact';
